requirements.txt
gradio==5.12.0
openai==1.59.8
Pillow==11.1.0
Requests==2.32.3
